+++ tests/testthat/test-vdiffr_check_model.R
@@ -246,3 +246,25 @@ test_that("ppc_range works", {
     fig = plot(performance::check_model(quine.nb1, x_limits = c(0, 10)))
   )
 })
+
+test_that("check_model() with heterogeneity plots works", {
+  set.seed(1)
+  n <- 600
+  size <- 20
+  x <- runif(n, -3, 3)
+  d <- data.frame(x = x, y = rbinom(n, size, plogis(-0.5 + 1.2 * x)))
+  d$f <- size - d$y
+
+  m <- glm(cbind(y, f) ~ x, family = binomial, data = d)
+  expect_doppelganger_with_seed(
+    title = "check_model_heterogeneity-1",
+    fig = plot(performance::check_model(m))
+  )
+
+  skip_if_not_installed("glmmTMB")
+  m <- glmmTMB::glmmTMB(cbind(y, f) ~ x, family = binomial, data = d)
+  expect_doppelganger_with_seed(
+    title = "check_model_heterogeneity-2",
+    fig = plot(performance::check_model(m))
+  )
+})